refactor(backend): adopt upstream PR #2613 interface changes (3 of 4)
Picks up bobtista's interface commits from TheSuperHackers/GeneralsGameCode
PR #2613 so our render backend interface matches upstream by construction
rather than by a parallel edit, and mirrors each into D3D11Backend.

Adopted:
  4d2aeeb0  shader ids to fixed-width unsigned int (unsigned long is 64-bit
            on targets where long is 64-bit)
  8acd5f0d  Set_Index_Buffer base offsets widened to unsigned int; the DX8
            forwards narrow with an explicit cast. Draw_Triangles/Draw_Strip
            were already unsigned int here, so only the index half applied.
  bdd6d938  drop the SurfaceClass back-buffer accessor - no caller anywhere
            in the tree reached it through the interface.

DX8Wrapper's own statics keep their unsigned short signatures; backends narrow
on the way through, per the interface contract.

Declined: 77b2768d (drop the Set_Gamma uselimit parameter). Upstream can drop
it because nothing in that branch passes anything but true. This tree has live
callers passing false - W3DDisplay.cpp:501 (Generals) and :582 (GeneralsMD) -
so taking it would silently apply gamma limiting where the engine asked for it
off. Reported on the PR.

Verified: Debug build of z_generals + w3d_d3d11_smoke exit 0; Release build
exit 0; w3d_d3d11_smoke 20/20 checks SMOKE PASS; frame_oracle selftest PASS
(negative control for the parity oracle).

+++ Core/Libraries/Source/WWVegas/WW3D2/Backend/D3D11Backend_W3D.cpp
@@ -510,7 +510,7 @@ void D3D11Backend::Set_Vertex_Buffer(const DynamicVBAccessClass & vba)
 	DX8Wrapper::Set_Vertex_Buffer(vba);
 }
 
-void D3D11Backend::Set_Index_Buffer(const IndexBufferClass * ib, unsigned short index_base_offset)
+void D3D11Backend::Set_Index_Buffer(const IndexBufferClass * ib, unsigned int index_base_offset)
 {
 	m_boundIndexBuffer = ib;
 	m_indexBaseOffset = index_base_offset;
@@ -526,10 +526,10 @@ void D3D11Backend::Set_Index_Buffer(const IndexBufferClass * ib, unsigned short 
 		}
 	}
 	// Mirror into the engine-side render_state record (see Set_Vertex_Buffer).
-	DX8Wrapper::Set_Index_Buffer(ib, index_base_offset);
+	DX8Wrapper::Set_Index_Buffer(ib, static_cast<unsigned short>(index_base_offset));
 }
 
-void D3D11Backend::Set_Index_Buffer(const DynamicIBAccessClass & iba, unsigned short index_base_offset)
+void D3D11Backend::Set_Index_Buffer(const DynamicIBAccessClass & iba, unsigned int index_base_offset)
 {
 	// The uploaded slice is 0-based within itself, so the DrawIndexed base-vertex
 	// is 0 for the dynamic path (the DX8 SetIndices base only applies to shared
@@ -548,7 +548,7 @@ void D3D11Backend::Set_Index_Buffer(const DynamicIBAccessClass & iba, unsigned s
 	// Mirror into the engine-side render_state record; the ENGINE-side
 	// index_base_offset is passed through unchanged (Flush uses it as a CPU-side
 	// vertex offset), only the GPU DrawIndexed base above stays 0.
-	DX8Wrapper::Set_Index_Buffer(iba, index_base_offset);
+	DX8Wrapper::Set_Index_Buffer(iba, static_cast<unsigned short>(index_base_offset));
 }
 
 // Real body here (not D3D11Backend.cpp) because the render_state mirror needs
